end-to-end-project_rainfall_prediction/requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
xgboost==1.7.5
zipp==3.15.0
